pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pythonproject"
version = "1.0.0"
description = "Python接口自动化测试平台"
authors = [
    {name = "bgtech", email = "bgtech@example.com"}
]
readme = "README.md"
requires-python = ">=3.8,<3.9"
license = {text = "MIT"}
keywords = ["testing", "api", "automation", "pytest"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Testing :: Acceptance",
    "Topic :: Software Development :: Testing :: BDD",
]

dependencies = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-html>=3.0.0,<4.0.0",
    "requests>=2.28.0,<3.0.0",
    "PyYAML>=6.0,<7.0.0",
    "pandas>=1.5.0,<2.0.0",
    "openpyxl>=3.0.0,<4.0.0",
    "pymysql>=1.0.0,<2.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "redis>=4.0.0,<5.0.0",
    "pika>=1.3.0,<2.0.0",
    "rocketmq-client-python>=2.0.0,<3.0.0",
    "typing-extensions>=4.0.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
python_version = ">=3.8,<3.9"
testpaths = ["testcase"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
] 